Backend: Add feature to limit the concurrent submission processing from SQS(#2129)

+++ scripts/workers/submission_worker.py
@@ -547,12 +547,24 @@ def main():
     queue_name = os.environ.get('CHALLENGE_QUEUE', 'evalai_submission_queue')
     queue = get_or_create_sqs_queue(queue_name)
 
+    maximum_concurrent_submissions = challenge.max_concurrent_submission_evaluation
     while True:
         for message in queue.receive_messages():
-            logger.info('Processing message body: {0}'.format(message.body))
-            process_submission_callback(message.body)
-            # Let the queue know that the message is processed
-            message.delete()
+            if settings.DEBUG or settings.TEST:
+                logger.info('Processing message body: {0}'.format(message.body))
+                process_submission_callback(message.body)
+                # Let the queue know that the message is processed
+                message.delete()
+            else:
+                current_running_submissions_count = Submission.objects.filter(
+                    challenge_phase__challenge=challenge.id, status='running').count()
+                if current_running_submissions_count == maximum_concurrent_submissions:
+                    pass
+                else:
+                    logger.info('Processing message body: {0}'.format(message.body))
+                    process_submission_callback(message.body)
+                    # Let the queue know that the message is processed
+                    message.delete()
         if killer.kill_now:
             break
         time.sleep(0.1)
